Add Filter tahun dan bulan

+++ resources/views/MasterData/GsmMaster/index.blade.php
@@ -57,6 +57,7 @@
                 <th scope="col" class="list">Active Date</th>
                 <th scope="col" class="list">Terminated Date</th>
                 <th scope="col" class="list">Note</th>
+                <th scope="col" class="list">Provider</th>
                 <th scope="col" class="action">Action</th>
               </tr>
             </thead>
@@ -110,6 +111,7 @@
                       <th>Active Date</th>
                       <th>Terminated Date</th>
                       <th>Note</th>
+                      <th>Provider</th>
                     </tr>
                   </thead>
                   <tbody id="item_data_temporary">
@@ -146,6 +148,10 @@
     // ------- send import -----
     function send_data() {
       $rowCount = $("#table_temporary_id tr").length;
+      if ($rowCount == 1) {
+        alert('table empty')
+      } else {
+
       $rowResult = $rowCount - 1;
       var allVals = [];
       for($i = 0; $i < $rowResult; $i++)
@@ -167,37 +173,46 @@
         var active_date = $(".temporary-active_date-"+value).attr("id");
         var terminate_date = $(".temporary-terminate_date-"+value).attr("id");
         var note = $(".temporary-note-"+value).attr("id");
-          $.ajax({
-          type: "get",
-          url: "{{ url('store_GsmMaster') }}",
-          data: {
-            status_gsm: status_gsm,
-            gsm_number: gsm_number,
-            company_id: company_id,
-            serial_number: serial_number,
-            icc_id: icc_id,
-            imsi: imsi,
-            res_id: res_id,
-            request_date: request_date,
-            expired_date: expired_date,
-            active_date: active_date,
-            terminate_date: terminate_date,
-            note: note
-          },
-          success: function(data) {
-            swal({
-              type: 'success',
-              title: 'Data Saved',
-              showConfirmButton: false,
-              timer: 1500
-          }).catch(function(timeout) { });
-            read();
-            deleteTemporary();
-            read_temporary()
-            $('#importData').modal('hide');
+        var provider = $(".temporary-provider-"+value).attr("id");
+          if ( request_date == null) {
+            alert('column is empty')
+          } else {
+            $.ajax({
+            type: "get",
+            url: "{{ url('store_GsmMaster') }}",
+            data: {
+              status_gsm: status_gsm,
+              gsm_number: gsm_number,
+              company_id: company_id,
+              serial_number: serial_number,
+              icc_id: icc_id,
+              imsi: imsi,
+              res_id: res_id,
+              request_date: request_date,
+              expired_date: expired_date,
+              active_date: active_date,
+              terminate_date: terminate_date,
+              note: note,
+              provider: provider
+            },
+            success: function(data) {
+                swal({
+                  type: 'success',
+                  title: 'Data Saved',
+                  showConfirmButton: false,
+                  timer: 1500
+              }).catch(function(timeout) { });
+                read();
+                deleteTemporary();
+                read_temporary()
+                $('#importData').modal('hide');
+              }
+          });
           }
       });
-      });
+        
+      }
+      
     }
 
     // ---- Close Modal -------
@@ -293,6 +308,7 @@
         var active_date = $("#active_date").val();
         var terminate_date = $("#terminate_date").val();
         var note = $("#note").val();
+        var provider = $("#provider").val();
         $.ajax({
             type: "get",
             url: "{{ url('store_GsmMaster') }}",
@@ -308,7 +324,8 @@
               expired_date: expired_date,
               active_date: active_date,
               terminate_date: terminate_date,
-              note: note
+              note: note,
+              provider: provider
             },
             success: function(data) {
              swal({
@@ -372,6 +389,7 @@
         $("#item-active_date-"+id).hide("fast");
         $("#item-terminate_date-"+id).hide("fast");
         $("#item-note-"+id).hide("fast");
+        $("#item-provider-"+id).hide("fast");
         $.get("{{ url('show_GsmMaster') }}/" + id, {}, function(data, status) {
             $("#edit-form-"+id).prepend(data)
         });
@@ -390,6 +408,7 @@
           var active_date = $("#active_date").val();
           var terminate_date = $("#terminate_date").val();
           var note = $("#note").val();
+          var provider = $("#provider").val();
           var id = id;
             $.ajax({
                 type: "get",
@@ -406,7 +425,8 @@
                   expired_date: expired_date,
                   active_date: active_date,
                   terminate_date: terminate_date,
-                  note: note
+                  note: note,
+                  provider: provider
                 },
                 success: function(data) {
                 swal({
@@ -507,6 +527,7 @@
                     $("#item-active_date-"+value).hide("fast");
                     $("#item-terminate_date-"+value).hide("fast");
                     $("#item-note-"+value).hide("fast");
+                    $("#item-provider-"+value).hide("fast");
                     $(".add").hide("fast");
                     $.get("{{ url('show_GsmMaster') }}/" + value, {}, function(data, status) {
                         $("#edit-form-"+value).prepend(data)
@@ -548,6 +569,7 @@
                     var active_date = $(".active_date-"+value).val();
                     var terminate_date = $(".terminate_date-"+value).val();
                     var note = $(".note-"+value).val();
+                    var provider = $(".provider-"+value).val();
                     $.ajax({
                     type: "get",
                     url: "{{ url('update_GsmMaster') }}/"+value,
@@ -563,7 +585,8 @@
                       expired_date: expired_date,
                       active_date: active_date,
                       terminate_date: terminate_date,
-                      note: note
+                      note: note,
+                      provider: provider
                     },
                     success: function(data) {
                      swal({
